PowerCrust algo.
The powerCrust algorithm:

As Input the algorithm wants a point cloud in either 2D or 3D (3 or 2 by how many samples you have) and will output a surface estimation trough a mesh.  The option to plot is present in the code, which will output the Voronoi diagram and powerdigram. Its easiest to see these plots in 2D.  The code also has a special function named STLTest, which gives the user possibility to import two different breast models.